avoid special characters in slug header test
Per RFC 5023, the server may change or remove special characters in the
Slug header when creating resources. Try to pick a string that will
likely be accepted as-is (a short string with only alphabetic
characters). Also ignore case when checking if the returned location
contains the slug.

See http://tools.ietf.org/html/rfc5023#page-30